+++ CHANGELOG.md
@@ -575,6 +575,7 @@
 
 * Changed SQL persistence to work with `IDbConnectionProvider` as originally intended - thanks [MrMDavidson]
 * Made Azure Service Bus transport use the native message deferral mechanism
+* Made Azure Storage Queues transport use the native message deferral mechanism
 
 
 [AndreaCuneo]: https://github.com/AndreaCuneo